Milestone: Projects 1 to 3 (RockPaperScissors)
This is an implementation of the milestone change which was to create a small game app based on Rock Paper Scissors. The game makes a random selection each round then challenges the player to either select an option that would win the game or lose the game. In my version, players receive 10 points if they choose correctly or lose 5 points if they choose incorrectly. The main view uses custom properties to provide the various view parts and a custom VStack that includes some font styling relevant to the game's UI. I couldn't resist using some emojis to represent the three choices.